% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot-rbcc}
\alias{plot-rbcc}
\alias{plot.rbcc}
\alias{plot.rbcusumcc}
\alias{plot.rbmcc}
\title{Plot Risk-Based Control Charts}
\usage{
\method{plot}{rbcc}(x, title = NULL, xlab = "Group", ylab = "Chart statistic", ...)

\method{plot}{rbcusumcc}(x, title = NULL, xlab = "Group", ylab = "Cumulative sum", ...)

\method{plot}{rbmcc}(x, title = NULL, xlab = "Group", ylab = "T-squared statistic", ...)
}
\arguments{
\item{x}{A fitted risk-based chart object.}

\item{title}{Optional title.}

\item{xlab, ylab}{Axis labels.}

\item{...}{Reserved for future use.}
}
\value{
A visible \code{ggplot} object.
}
\description{
Plot Risk-Based Control Charts
}
