% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print-rbcc}
\alias{print-rbcc}
\alias{print.rbcc}
\alias{print.rbcusumcc}
\alias{print.rbmcc}
\alias{print.summary.rbcc}
\alias{print.summary.rbcusumcc}
\alias{print.summary.rbmcc}
\title{Print Risk-Based Control-Chart Objects}
\usage{
\method{print}{rbcc}(x, digits = getOption("digits"), ...)

\method{print}{rbcusumcc}(x, digits = getOption("digits"), ...)

\method{print}{rbmcc}(x, digits = getOption("digits"), ...)

\method{print}{summary.rbcc}(x, digits = getOption("digits"), ...)

\method{print}{summary.rbcusumcc}(x, digits = getOption("digits"), ...)

\method{print}{summary.rbmcc}(x, digits = getOption("digits"), ...)
}
\arguments{
\item{x}{A fitted or summarized chart object.}

\item{digits}{Number of significant digits.}

\item{...}{Additional print arguments.}
}
\value{
The input object, invisibly.
}
\description{
Print Risk-Based Control-Chart Objects
}
