\name{rbcc-package}
\alias{rbcc-package}
\docType{package}

\title{Package of Risk-based Control Charts}

\description{The risk-based is a new methodology to design an optimized control chart that minimized the cost of decision outcomes of the control process. The basic purpose of the risk-based control is to determine the optimal control charts parameters to minimize the risks arising from measurement uncertainty. This article develops an R package for family of risk-based control charts, namely \code{'rbcc'}. In this package, the functions required in the design of family of risk-based control charts univariate and multivariate

This work has been implemented by the TKP2021-NVA-10 project with the support provided by the Ministry of Culture and Innovation of Hungary from the National Research, Development and Innovation Fund, financed under the 2021 Thematic Excellence Programme funding scheme.
}

\seealso{\code{\link{rbcc}}, \code{\link{data_gen}}, \code{\link{summary.rbcc}}, \code{\link{plot.rbcc}}.}

\author{Aamir Saghir, Attila I. Katona, Zsolt T. Kosztyan*

e-mail: kzst@gtk.uni-pannon.hu}

\references{
Katona, A. I., Saghir, A., Hegedűs, C., & Kosztyán, Z. T. (2023). Design of Risk-Based Univariate Control Charts with Measurement Uncertainty. IEEE Access, 11, 97567-97573.

Katona, A. I. (2021). Validation of risk-based quality control techniques: a case study from the automotive industry. Journal of Applied Statistics, 1-20.

Kosztyán, Z. T., & Katona, A. I. (2016). Risk-based multivariate control chart. Expert Systems with Applications, 62, 250-262.
}

\keyword{package}
