% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{rbcc_control}
\alias{rbcc_control}
\title{Construct Optimization Controls}
\usage{
rbcc_control(grid_size = 201L, tol = .Machine$double.eps^0.25)
}
\arguments{
\item{grid_size}{Number of points used by the grid optimizer.}

\item{tol}{Positive numerical tolerance.}
}
\value{
A list suitable for the \code{control} argument.
}
\description{
Construct Optimization Controls
}
