% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary-rbcc}
\alias{summary-rbcc}
\alias{summary.rbcc}
\alias{summary.rbcusumcc}
\alias{summary.rbmcc}
\title{Summarize Risk-Based Control-Chart Objects}
\usage{
\method{summary}{rbcc}(object, ...)

\method{summary}{rbcusumcc}(object, ...)

\method{summary}{rbmcc}(object, ...)
}
\arguments{
\item{object}{A fitted risk-based chart object.}

\item{...}{Reserved for future use.}
}
\value{
A structured summary object retaining all original components.
}
\description{
Summarize Risk-Based Control-Chart Objects
}
