\name{rbcc-package}
\alias{rbcc-package}
\docType{package}
\title{Risk-Based Control Charts}
\description{
Tools for risk-based statistical process control in the presence of
measurement uncertainty. The package evaluates four decision outcomes and
optimizes chart parameters by minimizing total decision cost.
}
\details{
Available charts include X-bar, R, S, moving-average, EWMA, CUSUM, and
multivariate T-squared charts. See \code{citation("rbcc")} for references.
}
\author{Aamir Saghir, Attila Imre Katona, and Zsolt Tibor Kosztyan.}
\keyword{package}
