\name{rbcc}
\alias{rbcc}
\docType{package}

\title{Package of Risk-based Control Charts}

\description{
Risk-based Multivariate control charts.
}

\seealso{\code{\link{rbmcc}}, \code{\link{rbmcc_opt}}, \code{\link{data_gen}}, \code{\link{plot.rbmcc}}, \code{\link{summary.rbmcc}}.}

\author{Aamir Saghir, Attila I. Katona, Zsolt T. Kosztyan*

e-mail: kzst@gtk.uni-pannon.hu}

\references{
Kosztyán, Z. T., and Katona, A. I. (2016). Risk-based multivariate control chart. Expert Systems with Applications, 62, 250-262.
}

\keyword{package}
